Move sort_by_cluster in clustering
